Updated CMakeLists.txt so suffix does not affect other platforms.

+++ src/ChromeHost/CmdMute.cpp
@@ -45,7 +45,7 @@ bool CmdMute::CanExecute(std::string cmd)
 
 void CmdMute::Execute(std::string cmd)
 {
-  unsigned deviceId = m_headsetIntegrationService->GetCurrentDeviceId();
+  unsigned short deviceId = m_headsetIntegrationService->GetCurrentDeviceId();
   if (deviceId == USHRT_MAX)
   {
     m_headsetIntegrationService->Error("No device");
