+++ src/ChromeHost/CmdOnHook.cpp
@@ -45,7 +45,7 @@ bool CmdOnHook::CanExecute(std::string cmd)
 
 void CmdOnHook::Execute(std::string cmd)
 {
-  unsigned deviceId = m_headsetIntegrationService->GetCurrentDeviceId();
+  unsigned short deviceId = m_headsetIntegrationService->GetCurrentDeviceId();
   if (deviceId == USHRT_MAX)
   {
     m_headsetIntegrationService->Error("No device");
